pyproject.toml
[project]
name = "project-2-working-with-airflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow==2.11.0",
]
